Refactor: refactor question api directory
